requirements.txt
# Python
python-dotenv~=1.0.0
PyJWT~=2.7.0

# Django
Django==4.2.1
djangorestframework
djangorestframework-simplejwt==5.2.2
django-filter==23.2

# Database
psycopg2-binary
dj_database_url==2.0.0

# Security
django-cors-headers

# Deployment
gunicorn
uvicorn[standard]==0.22.0

# Distributed async tasks
celery
django-celery-beat

# Channels
channels
channels_redis

# Redis - needed both for channels and celery
django-redis==5.2.0

# Documentation
drf-spectacular==0.26.2
